renaming to selected instead of filtered
