+++ web/libs/datamanager/src/components/CellViews/AgreementSelected.jsx
