+++ web/libs/datamanager/src/components/CellViews/index.js
@@ -1,7 +1,7 @@
 import { toStudlyCaps } from "@humansignal/core";
 
 export { Agreement } from "./Agreement/Agreement";
-export { AgreementFiltered } from "./AgreementFiltered";
+export { AgreementSelected } from "./AgreementSelected";
 export {
   Annotators,
   Annotators as Reviewers,
