+++ web/libs/datamanager/src/components/Filters/types/index.js
@@ -3,5 +3,5 @@ export { Common } from "./Common";
 export { DateFilter as Date } from "./Date";
 export { DatetimeFilter as Datetime } from "./Datetime";
 export { ListFilter as List } from "./List";
-export { NumberFilter as Number, NumberFilter as AgreementFiltered } from "./Number";
+export { NumberFilter as Number, NumberFilter as AgreementSelected } from "./Number";
 export { StringFilter as Image, StringFilter as String } from "./String";
